requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.2
gTTS==2.5.4
idna==3.11
jieba3k==0.35.1
joblib==1.5.3
lxml==6.0.2
newspaper3k==0.2.8
nltk==3.9.2
pillow==12.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.1
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
